finalll

+++ models/staging/raw/stg_raw__product.sql
@@ -10,7 +10,7 @@ renamed as (
 
     select
         products_id,
-      CAST (purchse_price AS FLOAT) as purchase_price
+      CAST (purchse_price AS FLOAT64) as purchase_price
 
     from source
